Move Account component to containers directory and fix imports/exports
